+++ SOURCE/encode_qc.h
@@ -1,6 +1,6 @@
 #ifndef _ENCODE_QC_H_
 #define _ENCODE_QC_H_
 
-void encode_qc( int codeword[], short *hb[], int M, int b, int c, short synd[] );
+void encode_qc( int codeword[], int *hb[], int M, int b, int c, int synd[] );
 
 #endif //_ENCODE_QC_H_
